Split test_storage.rb into tets for each storage
