+++ test/storage/test_storage_error.rb
